refactor: modularize VaDER;
