--- pypots/clustering/vader/modules/__init__.py
@@ -1,14 +1,0 @@
-"""
-
-"""
-
-# Created by Wenjie Du <wenjay.du@gmail.com>
-# License: BSD-3-Clause
-
-from .core import _VaDER
-from .submodules import inverse_softplus
-
-__all__ = [
-    "_VaDER",
-    "inverse_softplus",
-]
